requirements.txt
composabl
flask[async]
numpy
Jinja2
paho-mqtt